+++ NEWS.md
@@ -1,9 +1,13 @@
-# MiscMetabar 0.13.0 (in development) 
+# MiscMetabar 0.13.1 (in development) 
+
+- Better filter of parameters in [add_new_taxonomy_pq()]. Only parameters used by the assign_* function corresponding to `method` are used.
+
+# MiscMetabar 0.13.0
 
 - Add function [assign_blastn()] and add a method `blast` in the function [add_new_taxonomy_pq()].
 - Add function [resolve_vector_ranks()] to resolve conflict in a vector of taxonomy values
 
-# MiscMetabar 0.12.1 (in development) 
+# MiscMetabar 0.12.1
 
 - Add parameter name `min_bootstrap` in [add_new_taxonomy_pq()]
 - Bug fix in [assign_idtaxa()]